refactor(core): Rename AddNewChartWorkflow to EditChartDetailsWorkflow
