+++ app/src/main/kotlin/com/omricat/maplibrarian/chartlist/ChartsViewFactories.kt
@@ -22,6 +22,6 @@ internal val MapListViewRegistry = ViewRegistry(
     ChartListLoadingViewFactory,
     ChartErrorViewFactory,
     ChartListLayoutRunner,
-    AddItemScreenViewFactory,
+    EditChartDetailsScreenViewFactory,
     SavingItemScreenViewFactory
 )
